+++ src/app/admin/category/category.component.ts
@@ -2,7 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ICategoryResponse } from 'src/app/shared/interfaces/category/category.interface';
 import { CategoryService } from 'src/app/shared/services/category/category.service';
-import { Storage, deleteObject, getDownloadURL, percentage, ref, uploadBytesResumable } from '@angular/fire/storage';
+import { ToastrService } from 'ngx-toastr';
+import { ImageService } from 'src/app/shared/services/image/image.service';
 
 
 @Component({
@@ -23,7 +24,8 @@ export class CategoryComponent implements OnInit {
   constructor(
     private fb: FormBuilder,
     private categoryService: CategoryService,
-    private storage: Storage
+    private imageService: ImageService,
+    private toastr: ToastrService
   ) { }
 
   ngOnInit(): void {
@@ -48,16 +50,21 @@ export class CategoryComponent implements OnInit {
   changeStatusForAddNewCategory(): void {
     this.addNewCategory = !this.addNewCategory;
   }
+
   saveCategory(): void {
     if (this.editStatus) {
       this.categoryService
         .updateCategory(this.categoryForm.value, this.currentCategoryId)
         .subscribe(() => {
           this.loadCategories();
+          this.toastr.success('Category successfully updated');
+
         });
     } else {
       this.categoryService.createCategory(this.categoryForm.value).subscribe(() => {
         this.loadCategories();
+        this.toastr.success('Category successfully created');
+
       });
     }
     this.isUploaded = false;
@@ -79,14 +86,18 @@ export class CategoryComponent implements OnInit {
   }
 
   deleteCategory(category: ICategoryResponse): void {
-    this.categoryService.deleteCategory(category.id).subscribe(() => {
-      this.loadCategories();
-    })
+    if (confirm(`Do you want to delete category ${category.name}`)) {
+      this.categoryService.deleteCategory(category.id).subscribe(() => {
+        this.loadCategories();
+        this.toastr.success('Category successfully deleted');
+
+      })
+    }
   }
 
   upload(event: any): void {
     const file = event.target.files[0];
-    this.uploadFile('category-images', file.name, file)
+    this.imageService.uploadFile('category-images', file.name, file)
       .then(data => {
         this.categoryForm.patchValue({
           imagePath: data
@@ -98,37 +109,16 @@ export class CategoryComponent implements OnInit {
       })
   }
 
-  async uploadFile(folder: string, name: string, file: File | null): Promise<string> {
-    const path = `${folder}/${name}`;
-    let url = '';
-    if(file) {
-      try {
-        const storageRef = ref(this.storage, path);
-        const task = uploadBytesResumable(storageRef, file);
-        percentage(task).subscribe(data => {
-          this.uploadPercent = data.progress
-        });
-        await task;
-        url = await getDownloadURL(storageRef);
-      } catch (e: any) {
-        console.error(e);
-      }
-    } else {
-      console.log('wrong format');
-    }
-    return Promise.resolve(url);
-  }
-
   deleteImage(): void {
-    const task = ref(this.storage, this.valueByControl('imagePath'));
-    deleteObject(task).then(() => {
-      console.log('File deleted');
-      this.isUploaded = false;
-      this.uploadPercent = 0;
-      this.categoryForm.patchValue({
-        imagePath: null
+    this.imageService.deleteUploadFile(this.valueByControl('imagePath'))
+      .then(() => {
+        this.isUploaded = false;
+        this.uploadPercent = 0;
+        this.categoryForm.patchValue({ imagePath: null });
       })
-    })
+      .catch(err => {
+        console.log(err);
+      })
   }
 
   valueByControl(control: string): string {
